added individual files login and addCourse validation
